fix: include model_spec_types module missing from P06 commit
